# feat(WAProto): Optimize protobuf generation for 80%+ bundle size reduction