--- a/src/Socket/groups.ts
+++ b/src/Socket/groups.ts
@@ -239,7 +239,7 @@ export const makeGroupsSocket = (config: SocketConfig) => {
 				// update the invite message to be expired
 				if (key.id) {
 					// create new invite message that is expired
-					inviteMessage = proto.Message.GroupInviteMessage.fromObject(inviteMessage)
+					inviteMessage = proto.Message.GroupInviteMessage.create(inviteMessage)
 					inviteMessage.inviteExpiration = 0
 					inviteMessage.inviteCode = ''
 					ev.emit('messages.update', [