… WAProto/index.d.ts
… WAProto/index.js