Optimization to the protobuf asset generation process, resulting in a dramatic reduction in the final bundle size of `index.js` and `index.d.ts`. The generated `index.js` file has been reduced from **7.9MB to 1.5MB** (>80% reduction), which will improve package load times, CI/CD speed, and overall application performance.

### The Problem

The previous protobuf generation script produced very large files containing heavyweight conversion and verification methods (`fromObject`, `toJSON`, etc.) and non-minified code. These utilities, while useful for ensuring type safety with untrusted data, add significant overhead and were largely unnecessary for our internal use case.

### The Solution

This was achieved through a two-pronged approach:

1.  **Build Script Optimization (`GenerateStatics.sh`)**:
    *   The `pbjs` command was updated with the `--no-convert` flag to prevent the generation of the heavy `.fromObject()`, `.toObject()`, and `.toJSON()` methods.
    *   The `--no-beautify` flag was added to minify the output, removing unnecessary whitespace.

2.  **Source Code Adaptation**:
    *   To accommodate the removal of the conversion methods, all instances of `proto.Message.fromObject(obj)` have been replaced with the more lightweight and performant `proto.Message.create(obj)`. The `create` method is perfectly suited for our use case, where we are constructing objects that are already known to have the correct structure.
    *   Calls to `.toJSON()` after `.decode()` were removed, as `.decode()` now returns a plain object directly.

### Impact & Results

The size reduction of the generated assets is substantial:

**Before:**
```
WAProto
├── [2.3M]  index.d.ts
├── [7.9M]  index.js
└── [146K]  WAProto.proto
```

**After:**
```
WAProto
├── [525K]  index.d.ts  (-77%)
├── [1.5M]  index.js  (-81%)
└── [146K]  WAProto.proto
```

Running yarn tsc (type checking):
| Branch | Elapsed Time | User Time (CPU) | Max Memory |
| :--- | :---: | :---: | :---: |
| `master` | 3.48s | 8.08s | 390 MB |
| **`optimize-proto`** | **3.09s** | **7.08s** | **371 MB** |
| **Improvement** | **-11.2%** | **-12.4%** | **-4.9%** |

### Verification

All necessary code changes to support the new, leaner protobuf asset have been made. The project compiles successfully with `yarn tsc`, confirming that all type signatures are correct and the new methods are being used properly. There should be no functional regressions from this change.