@purpshell 2025-08-06
Nice change, but it is indeed a breaking one with the proto methods.. 
We'll have to let some sort of warning out or release this in a minor version