@joaosouz4dev 2025-08-13
> Nice change, but it is indeed a breaking one with the proto methods.. We'll have to let some sort of warning out or release this in a minor version

Why don't we create a development branch?